Renaming the "X-Forwaded-For" header
Renaming the `HTTP_X_FORWARDED_FOR` header to match the
"non-standard-but-widely-used" way to name it
(`X-Forwaded-For`).

Moreover, it was never been sent with the HTTP
request (cURL ignores all HTTP headers starting with `HTTP_`), so I
guess that name was just a confusion from the fact that PHP prefixes all
the HTTP headers by `HTTP_` in its `$_SERVER` variable.

It can also be
noted that this header might become obsolete since RFC7239 defines a new
`Forwaded` header:
https://tools.ietf.org/html/rfc7239